Delete basic data types. Add asteroid, bullet, player, game object. Fix window get size and resize.

+++ CS228 PROJECT/CS228 PROJECT/Input.cpp
@@ -119,7 +119,7 @@ void Input::SetReleasedMouse(sf::Mouse::Button mouseButton)
 void Input::SetMousePosition(sf::Vector2i position)
 {
     mMouse.x = position.x - Engine::GetWindow().GetSize().x / 2;
-    mMouse.y = -(position.y - Engine::GetWindow().GetSize().y / 2);
+    mMouse.y = -(position.y - static_cast<int>(Engine::GetWindow().GetSize().y) / 2);
 }
 
 void Input::Reset()
